semantic_search/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
PyYAML==6.0.1
pandas==2.1.1
sentence-transformers==2.2.2
torch==2.0.1
networkx==3.1
scikit-learn==1.3.0
